Lifted all of the test code for interpreter one level
