+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/components/suitability-choice-page-base.component.spec.ts
@@ -40,4 +40,41 @@ const configureTestBedFor = <T extends SuitabilityChoicePageBaseComponent>
   return new SuitabilityChoicePageBaseFixture(fixture);
 };
 
-export { configureTestBedFor as ConfigureTestBedForPageComponent, SuitabilityChoicePageBaseFixture };
+/**
+ * Base test for any yes/no true/false radio button screen
+ * Tests so that the html is bound in such a way that the user cannot proceed before selecting one choice.
+ * Generalising the test in one place such as with this, allows us to ensure that all pages follow a similar pattern.
+ * @param component The component to be tested.
+ */
+const cannotProceedUntilChoiceIsSelected = <T extends SuitabilityChoicePageBaseComponent>(component: Type<T>):
+  SuitabilityChoicePageBaseFixture<T> => {
+  const fixture = configureTestBedFor(component);
+
+  // when
+  fixture.submitIsClicked();
+
+  // then
+  expect(fixture.component.isFormInvalid).toBeTruthy();
+
+  // expect form to be erronous
+  const formContainer = fixture.debugElementByCss('#form-container');
+  expect(formContainer.classes['govuk-form-group--error']).toBeTruthy();
+
+  // and error message to be displayed
+  const errorMessage = fixture.debugElementByCss('#error-message');
+  expect(errorMessage.nativeElement).toBeTruthy();
+
+  // when selecting a radio button
+  fixture.radioBoxIsClicked('#choice-yes');
+
+  // then
+  fixture.submitIsClicked();
+
+  return fixture;
+};
+
+export {
+  configureTestBedFor as ConfigureTestBedForPageComponent,
+  SuitabilityChoicePageBaseFixture,
+  cannotProceedUntilChoiceIsSelected as CannotProceeedUntilChoiceIsSelected
+};
