+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/about-you/about-you.component.html
@@ -7,7 +7,7 @@
     <form [formGroup]="form" (ngSubmit)="continue()">
 
       <div id="form-container" [ngClass]="isFormInvalid ? 'govuk-form-group--error' : 'govuk-form-group'">
-        <fieldset class="govuk-fieldset" aria-describedby="ability-hint ability-error">
+        <fieldset class="govuk-fieldset">
           <div *ngIf="isFormInvalid">
             <span class="govuk-error-message">
               Please answer this question
@@ -44,8 +44,7 @@
                             Please provide more information
                           </span>
                         </div>
-                        <textarea formControlName="textInput" class="txtMore vh-text-width" id="moreDetails" name="moreDetails" rows="5" cols="80"
-                          aria-describedby="more-detail-hint" [ngClass]="isTextInputInvalid ? 'govuk-textarea--error' : 'govuk-textarea'"></textarea>
+                        <textarea formControlName="textInput" class="txtMore vh-text-width" id="details" rows="5" cols="80" [ngClass]="isTextInputInvalid ? 'govuk-textarea--error' : 'govuk-textarea'"></textarea>
                       </div>
                   </div>
