+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/interpreter/interpreter.component.spec.ts
@@ -1,31 +1,11 @@
 import { InterpreterComponent } from './interpreter.component';
-import { ConfigureTestBedForPageComponent } from '../../components/suitability-choice-page-base.component.spec';
+import { CannotProceeedUntilChoiceIsSelected } from '../../components/suitability-choice-page-base.component.spec';
 
 describe('InterpreterComponent', () => {
   it('cannot proceed to next step until pressing choice, after submit value is bound', () => {
-    const fixture = ConfigureTestBedForPageComponent(InterpreterComponent);
-    const component = fixture.component;
+    const fixture = CannotProceeedUntilChoiceIsSelected(InterpreterComponent);
 
-    // when
-    fixture.submitIsClicked();
-
-    // then
-    expect(component.isFormInvalid).toBeTruthy();
-
-    // expect form to be erronous
-    const formContainer = fixture.debugElementByCss('#form-container');
-    expect(formContainer.classes['govuk-form-group--error']).toBeTruthy();
-
-    // and error message to be displayed
-    const errorMessage = fixture.debugElementByCss('#error-message');
-    expect(errorMessage.nativeElement).toBeTruthy();
-
-    // when selecting a radio button
-    fixture.radioBoxIsClicked('#choice-yes');
-
-    // then
-    fixture.submitIsClicked();
-
-    expect(component.model.interpreter).toBe(true);
+    // and value is bound
+    expect(fixture.component.model.interpreter).toBe(true);
   });
 });
